chore: 유효성 검사하는 함수 -> util/common으로 이동
